+++ crates/cubed-application/src/ports/mod.rs
@@ -4,6 +4,7 @@ pub mod file_system;
 pub mod java_manager;
 pub mod port_manager;
 pub mod process_manager;
+pub mod resource_monitor;
 pub mod server_repository;
 
 pub use console::{ConsoleLine, ConsoleCallback, ConsoleManager};
@@ -12,4 +13,5 @@ pub use file_system::FileSystemManager;
 pub use java_manager::{JavaInstallation, JavaManager};
 pub use port_manager::PortManager;
 pub use process_manager::{ProcessInfo, ProcessManager};
+pub use resource_monitor::{ResourceMonitor, SystemStats, ServerStats};
 pub use server_repository::ServerRepository;